+++ src/config.js
@@ -25,6 +25,9 @@ module.exports = {
     // enforce a password for creating new sessions. set to null to disable
     password: 'sharkie4life',
 
+    // disable or enable localStorage sync (turn off if clients send over huge localStorage data, resulting in huge memory usages)
+    disableLocalStorageSync: false,
+
     //// REWRITE HEADER CONFIGURATION ////
 
     // removes reverse proxy headers