+++ src/util/addMoreErrorGuards.js
@@ -4,7 +4,8 @@
 const hGuard = require('testcafe-hammerhead/lib/request-pipeline/connection-reset-guard');
 const isConnectionResetError = hGuard.isConnectionResetError;
 hGuard.isConnectionResetError = function (err) {
-    if (isConnectionResetError(err) || err.code === 'ERR_INVALID_PROTOCOL' || err.code === 'ETIMEDOUT') {
+    // for some reason, ECONNRESET isn't handled correctly
+    if (isConnectionResetError(err) || err.code === 'ERR_INVALID_PROTOCOL' || err.code === 'ETIMEDOUT' || err.code === 'ECONNRESET') {
         return true;
     }
     console.error('Unknown crash-inducing error:', err);